+++ src/main/java/tdr/pet/weblibrary/service/impl/BookServiceImpl.java
@@ -60,21 +60,37 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void updateBookById(Long id, Book book) {
-        bookRepository.updateBookById(id, book);
+        if (bookRepository.existsById(id)) {
+            bookRepository.updateBookById(id, book);
+        } else {
+            throw new BookNotFoundException();
+        }
     }
 
     @Override
     public void updateBookByIsbn(String isbn, Book book) {
-        bookRepository.updateBookByIsbn(isbn, book);
+        if (bookRepository.existsByIsbn(isbn)) {
+            bookRepository.updateBookByIsbn(isbn, book);
+        } else {
+            throw new BookNotFoundException();
+        }
     }
 
     @Override
     public void deleteBookById(Long id) {
-        bookRepository.deleteById(id);
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+        } else {
+            throw new BookNotFoundException();
+        }
     }
 
     @Override
     public void deleteBookByISBN(String isbn) {
-        bookRepository.deleteBookByIsbn(isbn);
+        if (bookRepository.existsByIsbn(isbn)) {
+            bookRepository.deleteBookByIsbn(isbn);
+        } else {
+            throw new BookNotFoundException();
+        }
     }
 }
